// src/ComponentPatterns/interfaces/interfaces.ts
import {ReactElement} from "react";

export interface Product {
  id: string;
  title: string;
  img?: string;
}

export interface ProductContextProps {
  counter: number;
  increaseBy: (value:number) => void;
  product: Product;
  maxCount?: number;
}

export interface onChangeArgs {
  product: Product;
  count: number;
}

export interface ProductCardCardHandlers {
  count: number;
  isMaxCountReached: boolean;
  maxCount?: number;
  product: Product;

  increaseBy: (value:number) => void;
  reset: () => void;
}